Added Font class
